FEAT: Colors were set for each language
The color palette that was previously had has been improved. Now we have
the official "monokai" color palette with variations in some colors.

The most relevant improvement in this update is the custom coloring for
html, css, js and json. Each of these languages looks good with the
established color scheme.